requirements.txt
annotated-types==0.7.0
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
pika==1.3.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.16.0
typer==0.12.5
typing_extensions==4.12.2
